+++ docs/RELEASES.md
@@ -17,7 +17,7 @@ This is the list of project releases, past and current. To see tasks that are st
 ## <a name="10.0.1.0"></a>10.0.1.0 :sheep: <sup>`11`</sup>
 - [ ] Fix menu not opening
 - [ ] Add admin to nginx with production settings
-- [ ] Fix emails to remove email contact and tell how to recover and delete csv
+- [x] `230619>230619` Fix emails to remove email contact and tell how to recover and delete csv
 - [x] `230618>230619` Fix terms to remove email contact and tell how to recover and delete csv
 - [x] `230530>230617` Add link to purge csv into email with csv
 - [x] `230409>230409` Fix acceptance check bug for calling inside another transaction